Add CLI script
